remove templates from routing algorithms
