+++ include/engine/plugins/match.hpp
@@ -39,8 +39,8 @@ class MatchPlugin : public BasePlugin
 
   private:
     mutable SearchEngineData heaps;
-    mutable routing_algorithms::MapMatching<datafacade::BaseDataFacade> map_matching;
-    mutable routing_algorithms::ShortestPathRouting<datafacade::BaseDataFacade> shortest_path;
+    mutable routing_algorithms::MapMatching map_matching;
+    mutable routing_algorithms::ShortestPathRouting shortest_path;
     const int max_locations_map_matching;
 };
 }
