+++ include/engine/plugins/viaroute.hpp
@@ -29,10 +29,9 @@ class ViaRoutePlugin final : public BasePlugin
 {
   private:
     mutable SearchEngineData heaps;
-    mutable routing_algorithms::ShortestPathRouting<datafacade::BaseDataFacade> shortest_path;
-    mutable routing_algorithms::AlternativeRouting<datafacade::BaseDataFacade> alternative_path;
-    mutable routing_algorithms::DirectShortestPathRouting<datafacade::BaseDataFacade>
-        direct_shortest_path;
+    mutable routing_algorithms::ShortestPathRouting shortest_path;
+    mutable routing_algorithms::AlternativeRouting alternative_path;
+    mutable routing_algorithms::DirectShortestPathRouting direct_shortest_path;
     const int max_locations_viaroute;
 
   public:
