+++ src/engine/plugins/match.cpp
@@ -180,7 +180,7 @@ Status MatchPlugin::HandleRequest(const std::shared_ptr<const datafacade::BaseDa
     }
 
     // call the actual map matching
-    SubMatchingList sub_matchings = map_matching(*facade,
+    SubMatchingList sub_matchings = map_matching(facade,
                                                  candidates_lists,
                                                  parameters.coordinates,
                                                  parameters.timestamps,
@@ -211,7 +211,7 @@ Status MatchPlugin::HandleRequest(const std::shared_ptr<const datafacade::BaseDa
         // bi-directional
         // phantom nodes for possible uturns
         shortest_path(
-            *facade, sub_routes[index].segment_end_coordinates, {false}, sub_routes[index]);
+            facade, sub_routes[index].segment_end_coordinates, {false}, sub_routes[index]);
         BOOST_ASSERT(sub_routes[index].shortest_path_length != INVALID_EDGE_WEIGHT);
     }
