+++ src/engine/plugins/viaroute.cpp
@@ -88,16 +88,16 @@ Status ViaRoutePlugin::HandleRequest(const std::shared_ptr<const datafacade::Bas
     {
         if (route_parameters.alternatives && facade->GetCoreSize() == 0)
         {
-            alternative_path(*facade, raw_route.segment_end_coordinates.front(), raw_route);
+            alternative_path(facade, raw_route.segment_end_coordinates.front(), raw_route);
         }
         else
         {
-            direct_shortest_path(*facade, raw_route.segment_end_coordinates, raw_route);
+            direct_shortest_path(facade, raw_route.segment_end_coordinates, raw_route);
         }
     }
     else
     {
-        shortest_path(*facade,
+        shortest_path(facade,
                       raw_route.segment_end_coordinates,
                       route_parameters.continue_straight,
                       raw_route);
